requirements.txt
asgiref==3.2.4
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
gunicorn
django-heroku
django-solo==1.1.3
Pillow
